move BedrockChatModel out of converse package + modifs on LoggingInterceptor
