+++ langchain4j-bedrock/src/test/java/dev/langchain4j/model/bedrock/BedrockChatModelWithoutVisionIT.java
@@ -1,4 +1,4 @@
-package dev.langchain4j.model.bedrock.converse;
+package dev.langchain4j.model.bedrock;
 
 import static dev.langchain4j.model.bedrock.converse.TestedModels.AI_JAMBA_INSTRUCT;
 import static dev.langchain4j.model.bedrock.converse.TestedModels.AWS_NOVA_MICRO;
